refactor: split build_app_state into domain-specific builders (#685)
Extract the 718-line build_app_state into five focused builder functions,
each returning a typed *Bundle struct:

- builders/storage.rs  — dir validation, TaskStore, QValueStore
- builders/engines.rs  — RuleEngine, EventStore, GcAgent, SkillStore
- builders/registry.rs — ThreadDb, PlanDb+cache, ProjectRegistry,
                         WorkspaceManager, RuntimeStateStore
- builders/intake.rs   — TaskQueue, FeishuIntake, GitHub pollers,
                         QualityTrigger, completion callback
- builders/services.rs — interceptor stack, service layer,
                         RuntimeHostManager, snapshot restore

build_app_state becomes a ~100-line ordered orchestrator that documents
the dependency edges between phases via comments.

Each builder file includes focused unit tests (13 total):
storage (2), engines (2), registry (2), intake (1), services (1).
