+++ crates/harness-server/src/http/builders/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod engines;
+pub(crate) mod intake;
+pub(crate) mod registry;
+pub(crate) mod services;
+pub(crate) mod storage;
